% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorities.R
\name{priorities}
\alias{priorities}
\title{Daily priorities}
\usage{
priorities(free_time, studying, stay_up_late)
}
\arguments{
\item{free_time}{how much time (hours) is left after work}

\item{studying}{productive activities (hours)}

\item{stay_up_late}{extra hours added to the night}
}
\value{
enjoyable time
}
\description{
Daily priorities
}
